+++ stock_checkout_sync/README.rst
@@ -1,0 +1,131 @@
+==============================
+Stock Checkout Synchronization
+==============================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Alpha
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fwms-lightgray.png?logo=github
+    :target: https://github.com/OCA/wms/tree/13.0/stock_checkout_sync
+    :alt: OCA/wms
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/wms-13-0/wms-13-0-stock_checkout_sync
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/285/13.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
+
+Regroup moves at the same location when they reach an operation type.
+
+Example of use case:
+
+* We use several operation types for Picking Transfers (Pick Shelves, Pick Reserve)
+* We use several operation types for Packing (Pack, Pack Post)
+* The "Checkout Synchronization" flag is activated on the 2 Packing operation types
+* The Packing can occur in several areas (Packing/PACK1, Packing/PACK2, Packing/PACK3)
+
+Based on these assumptions, we have these pickings, all for the same order:
+
+* SHELF001, pick 2 goods in the shelves
+* RESER001, pick 2 goods in the reserve
+* PACK001, pack 3 of the goods (2 from SHELF001, 1 from RESER001)
+* PPOST001, pack the remaining good of the reserve
+
+The operator uses the button "Checkout Sync" on a pick transfer,
+for instance SHELF001, and select Packing/PACK2 as destination.
+We want all the lines that reach PACK001 to be updated with:
+
+* The move lines (except the one that goes to PPOST001) must now reach
+  Packing/PACK2
+* The same moves' destination must reach Packing/PACK2 as well:
+  we constrain the user to move the goods there as we want to move
+  all the goods at the same place for the packing.
+* As the moves' destinations are updated to the same packing location,
+  new move lines will have the correct destination as well
+
+
+.. figure:: https://raw.githubusercontent.com/OCA/wms/13.0/stock_checkout_sync/static/description/sync-assistant.png
+   :alt: Checkout Synchronization Assistant Screen
+   :width: 600 px
+
+
+Notes:
+
+Works best when used with ``stock_available_to_promise_release``:
+
+* When using the mentioned module, we assume that we release moves (which
+  creates the whole chain of moves) only when we know that we have the
+  quantities in stock (otherwise the module splits the delivery). Any exception
+  (which is now rare when you release this ways) that would lead to a backorder
+  is handled by canceling the remaining line.
+* When the module is not used, then the destination of the backorders may be
+  changed at the same time
+
+Compatible with ``stock_dynamic_routing``.
+
+.. IMPORTANT::
+   This is an alpha version, the data model and design can change at any time without warning.
+   Only for development or testing purpose, do not use in production.
+   `More details on development status <https://odoo-community.org/page/development-status>`_
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Configuration
+=============
+
+Activate syncing on the picking type form view (Inventory > Settings > Operations Types)
+by marking ``Checkout Synchronization``.
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/OCA/wms/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/wms/issues/new?body=module:%20stock_checkout_sync%0Aversion:%2013.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Camptocamp
+
+Contributors
+~~~~~~~~~~~~
+
+* Akim Juillerat <akim.juillerat@camptocamp.com>
+* Guewen Baconnier <guewen.baconnier@camptocamp.com>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
+
+.. image:: https://odoo-community.org/logo.png
+   :alt: Odoo Community Association
+   :target: https://odoo-community.org
+
+OCA, or the Odoo Community Association, is a nonprofit organization whose
+mission is to support the collaborative development of Odoo features and
+promote its widespread use.
+
+This module is part of the `OCA/wms <https://github.com/OCA/wms/tree/13.0/stock_checkout_sync>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
