+++ stock_checkout_sync/__manifest__.py
@@ -3,7 +3,7 @@
 {
     "name": "Stock Checkout Synchronization",
     "summary": "Sync location for Checkout operations",
-    "version": "13.0.1.0.0",
+    "version": "13.0.1.1.0",
     "development_status": "Alpha",
     "category": "Warehouse Management",
     "website": "https://github.com/OCA/wms",
